\name{albers01}
\alias{albers01}
\title{Josef Albers Homage to the Square}
\description{
  Josef Albers Homage to the Square adaptation.
}
\usage{
albers01(color = c("teal", "gray", "orangered"))
}
\arguments{
  \item{color}{ vector of three colors for the squares, in order from largest to smallest.
  }
}
\details{
  Example from the Los Angeles County Museum of Art.
}
\value{
  NULL
}
\references{}
\author{Stephen Kaluzny <skaluzny@tibco.com>}

\examples{
albers01()
}
\keyword{modart}
