\name{baer01}
\alias{baer01}
\title{Pale Yellow by Jo Baer}
\description{
  An image not unlike Pale Yellow (1964-65) b yjo Baer.
}
\usage{
baer01()
}
\value{
  NULL
}
\details{
  Based on a work in the Portland Art Museum.
}
\author{Stephen Kaluzny <skaluzny@tibco.com>}
\examples{
baer01()
}
\keyword{modart}
