\name{mondian0}
\alias{mondrian01}
\title{Draw Mondrain Art}
\description{
  Draws an example of Piet Mondrian's Composition art works.
}
\usage{
mondian01(nVLines = 4, nHLines = 3, lineWidth = 2, colLine = "black")
}
\arguments{
  \item{nVLines}{number of veritcal lines}
  \item{nHLines}{number of horizonatla lines}
  \item{lineWidth}{width of the lines}
  \item{colLine}{the color of the lines}
}
\value{
  NULL
}
\author{Stephen Kaluzny <skaluzny@tibco.com>}
\examples{
mondrian01()
}
\keyword{modart}
