\name{sander01}
\alias{sander01}
\title{Ludwig Sander's Pompeli XI}
\description{
  An image not unlike Pompeli XI by Ludwig Sander.
}
\usage{
sander01(sampleColors=TRUE,
    col = c("red", "violet", "#d70000", "purple", "darkred"))
}
\arguments{
  \item{sampleColors}{a logical value, if \code{TRUE},
    the colors are randomly selected from the \code{Reds} vector.
    Otherwise, the colors specified in \code{col} are used.
}
  \item{col}{a character vector of colors names for the polygons in the work
}
\details{
  Example of work from xyzzy.
}
\value{
  NULL
}
\references{ ~put references to the literature/web site here ~ }
\author{Stephen Kaluzny <skaluzny@tibco.com>}
}
\examples{
sanders01()
}
\keyword{modart}
